Add lan driver that creates a lan device and 3 children devices
Device online/offline is hooked up to refresh/off capabilities, as well
as a way to emit an event for an offline device via switch level with a
value < 80. Note use the cli or advanced web app to send cap commands
for an offline device. Used for testing device health of lan parent
child setups